Separation of streamlit and non-streamlit components for DallE
